Move API tests to better places
