+++ src/blockhash.rs
@@ -215,4 +215,17 @@ mod test {
         let h2 = BlockHash::from_str(&hs).unwrap();
         assert_eq!(h, h2);
     }
+
+    #[test]
+    fn to_from_string() {
+        let hex_hash = concat!(
+            "00000000000000000000000000000000",
+            "00000000000000000000000000000000",
+            "00000000000000000000000000000000",
+            "00000000000000000000000000000000",
+        );
+        let hash = BlockHash::from_str(hex_hash).unwrap();
+        let hash2 = hash;
+        assert_eq!(hash2.to_string(), hex_hash);
+    }
 }
