--- tests/api/blockhash.rs
@@ -1,28 +1,0 @@
-// Conserve backup system.
-// Copyright 2015, 2016, 2017, 2018, 2019, 2020 Martin Pool.
-
-// This program is free software; you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation; either version 2 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-
-use conserve::blockhash::BlockHash;
-use std::str::FromStr;
-
-#[test]
-fn to_from_string() {
-    let hex_hash = concat!(
-        "00000000000000000000000000000000",
-        "00000000000000000000000000000000",
-        "00000000000000000000000000000000",
-        "00000000000000000000000000000000",
-    );
-    let hash = BlockHash::from_str(hex_hash).unwrap();
-    let hash2 = hash;
-    assert_eq!(hash2.to_string(), hex_hash);
-}
