+++ tests/delete.rs
